泛化序列化 demo调试代码

+++ src/main/java/io/github/ketao1989/dubbo/api/IUserBiz.java
@@ -14,5 +14,5 @@ public interface IUserBiz {
      * @param id
      * @return
      */
-    public String queryName(int id);
+    public String queryName(long id);
 }
